+++ gradle/libs.versions.toml
@@ -9,10 +9,8 @@ test-ext = "1.1.5"
 junit = "4.13.2"
 truth = "1.0"
 espresso = "3.5.1"
-agp = "8.2.1"
+agp = "8.2.2"
 kotlin = "1.8.22"
-core-ktx = "1.12.0"
-material = "1.11.0"
 
 [libraries]
 recyclerview = { group = "androidx.recyclerview", name = "recyclerview", version.ref = "recyclerview" }
@@ -27,8 +25,6 @@ androidx-test-junit = { group = "androidx.test.ext", name = "junit", version.ref
 androidx-test-rules = { group = "androidx.test", name = "rules", version.ref = "test-rules" }
 espresso-core = { group = "androidx.test.espresso", name = "espresso-core", version.ref = "espresso" }
 espresso-contrib = { group = "androidx.test.espresso", name = "espresso-contrib", version.ref = "espresso" }
-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "core-ktx" }
-material = { group = "com.google.android.material", name = "material", version.ref = "material" }
 
 [plugins]
 androidApplication = { id = "com.android.application", version.ref = "agp" }
