+++ internal-test-support/src/main/AndroidManifest.xml
