+++ screencaptor-sample-tests/.gitignore
