+++ screencaptor-sample-tests/src/main/res/drawable/add_accounts.png
