+++ screencaptor-sample/src/main/AndroidManifest.xml
