+++ screencaptor-sample/src/main/res/drawable/add_accounts.png
