+++ screencaptor-sample/src/main/res/drawable/wf_logo.png
