+++ screencaptor-sample/src/main/res/layout/message_row_item.xml
