+++ screencaptor-sample/src/main/res/layout/screenshot_test.xml
