+++ screencaptor-sample/src/main/res/raw/dialog.png
